remove vision plugin
Before: attempt to render a page with the theming service, if fail
we use vision
Now: The theming service will never fail to render a page

The problem with using vision is that when you want to fall
back to the templates on disk and you're not in hapi-land
you have to do custom logic for that instance. Might as
well build that logic into the theming service and
everybody can render pages the exact same way.